refactor: Refactored some methods and paths
