+++ src/app/app.module.ts
@@ -4,11 +4,12 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { LoginUserRegularFormModule } from './login-user-regular-form';
-import { ChooseLoginModule } from './choose-login';
 import { RouterModule } from '@angular/router';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { LoginUserRegularFormModule } from './login-user-regular-form';
+import { ChooseLoginModule } from './choose-login';
 
 @NgModule({
   declarations: [
