+++ src/app/choose-login/choose-login.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ChooseLoginComponent } from './choose-login';
 import { RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 
+import { ChooseLoginComponent } from './index';
 
 @NgModule({
   declarations: [
