+++ src/app/choose-login/choose-login/choose-login.component.html
@@ -17,7 +17,7 @@
         <button
         (click)="routeRegularUser()"
           type="submit"
-          class="text-white bg-gradient-to-r from-green-400 via-green-500 to-green-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-green-300 dark:focus:ring-green-800 font-medium rounded-lg w-full mt-10 text-sm px-5 py-2.5 text-center mr-2 mb-2"
+          class="text-white bg-gradient-to-r from-blue-400 via-blue-500 to-blue-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800 font-medium rounded-lg w-full mt-10 text-sm px-5 py-2.5 text-center mr-2 mb-2"
         >
           REGULAR LOGIN
         </button>
@@ -27,7 +27,7 @@
         <button
         (click)="routeCorporateUser()"
           type="button"
-          class="text-white bg-gradient-to-r from-green-400 via-green-500 to-green-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-green-300 dark:focus:ring-green-800 font-medium rounded-lg w-full mt-10 text-sm px-5 py-2.5 text-center mr-2 mb-2"
+          class="text-white bg-gradient-to-r from-blue-700 via-blue-500 to-blue-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800 font-medium rounded-lg w-full mt-10 text-sm px-5 py-2.5 text-center mr-2 mb-2"
         >
           CORPORATE LOGIN
         </button>
