+++ src/app/create-user-regular-form/create-user-regular-form-component/index.ts
@@ -1,0 +1,2 @@
+export * from './create-user-regular-form-component';
+export * from './create-user-regular-form.service';
