+++ src/app/create-user-regular-form/index.ts
@@ -1,0 +1,2 @@
+export * from './create-user-regular-form-component';
+export * from './create-user-regular-form.module';
