+++ src/app/login-user-corporate-form/login-user-corporate-form.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { LoginUserCorporateFormComponent } from './login-user-corporate-form-component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+
+import { LoginUserCorporateFormComponent } from './index';
 
 
 @NgModule({
