+++ src/app/login-user-regular-form/index.ts
@@ -1,2 +1,2 @@
-export * from "./login-user-regular-form-component";
+export * from './login-user-regular-form-component';
 export * from './login-user-regular-form.module';
