+++ src/app/login-user-regular-form/login-user-regular-form-component/index.ts
@@ -1,1 +1,2 @@
 export * from './login-user-regular-form.component';
+export * from './login-user-regular-form.service';
