+++ src/app/login-user-regular-form/login-user-regular-form.module.ts
@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LoginUserRegularFormComponent } from './index';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
-
+import { LoginUserRegularFormComponent } from './index';
 
 @NgModule({
   declarations: [
